third_party/blink/web_tests/platform/webview/external/wpt/webxr/ar-module/idlharness.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of undefined (reading 'requestSession')"
[FAIL] XRSession interface: attribute environmentBlendMode
  assert_own_property: self does not have own property "XRSession" expected property "XRSession" missing
[FAIL] XRSession interface: attribute interactionMode
  assert_own_property: self does not have own property "XRSession" expected property "XRSession" missing
[FAIL] XRSession interface: xrSession must inherit property "environmentBlendMode" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrSession is not defined"
[FAIL] XRSession interface: xrSession must inherit property "interactionMode" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: xrSession is not defined"
[FAIL] XRView interface: attribute isFirstPersonObserver
  assert_own_property: self does not have own property "XRView" expected property "XRView" missing
Harness: the test ran to completion.
